MAINT: special: wrap `faddeeva`, `erf`, `log_ndtr` from xsf
[skip cirrus]
