+++ scipy/special/xsf_wrappers.cpp
@@ -8,6 +8,7 @@
 #include "xsf/cdflib.h"
 #include "xsf/digamma.h"
 #include "xsf/ellip.h"
+#include "xsf/erf.h"
 #include "xsf/expint.h"
 #include "xsf/fresnel.h"
 #include "xsf/gamma.h"
@@ -314,15 +315,12 @@ int xsf_sici(double x, double *si, double *ci) { return xsf::sici(x, si, ci); }
 int xsf_shichi(double x, double *si, double *ci) { return xsf::shichi(x, si, ci); }
 
 int xsf_csici(npy_cdouble x, npy_cdouble *si, npy_cdouble *ci) {
-    return xsf::sici(to_complex(x),
-		     reinterpret_cast<complex<double> *>(si),
-		     reinterpret_cast<complex<double> *>(ci));
+    return xsf::sici(to_complex(x), reinterpret_cast<complex<double> *>(si), reinterpret_cast<complex<double> *>(ci));
 }
 
 int xsf_cshichi(npy_cdouble x, npy_cdouble *shi, npy_cdouble *chi) {
-    return xsf::shichi(to_complex(x),
-		       reinterpret_cast<complex<double> *>(shi),
-		       reinterpret_cast<complex<double> *>(chi));
+    return xsf::shichi(to_complex(x), reinterpret_cast<complex<double> *>(shi),
+                       reinterpret_cast<complex<double> *>(chi));
 }
 
 double cephes__struve_asymp_large_z(double v, double z, Py_ssize_t is_h, double *err) {
@@ -382,8 +380,6 @@ double cephes_igam_fac(double a, double x) { return xsf::cephes::detail::igam_fa
 
 double cephes_lanczos_sum_expg_scaled(double x) { return xsf::cephes::lanczos_sum_expg_scaled(x); }
 
-double cephes_erf(double x) { return xsf::cephes::erf(x); }
-
 double cephes_erfc(double x) { return xsf::cephes::erfc(x); }
 
 double cephes_poch(double x, double m) { return xsf::cephes::poch(x, m); }
@@ -403,6 +399,30 @@ double cephes_expm1(double x) { return xsf::cephes::expm1(x); }
 double cephes_expn(int n, double x) { return xsf::cephes::expn(n, x); }
 
 double xsf_ellipe(double x) { return xsf::ellipe(x); }
+
+double xsf_erf(double x) { return xsf::erf(x); }
+
+npy_cdouble xsf_cerf(npy_cdouble z) { return to_ccomplex(xsf::erf(to_complex(z))); }
+
+double xsf_erfc(double x) { return xsf::erfc(x); }
+
+npy_cdouble xsf_cerfc(npy_cdouble z) { return to_ccomplex(xsf::erfc(to_complex(z))); }
+
+double xsf_erfcx(double x) { return xsf::erfcx(x); }
+
+npy_cdouble xsf_cerfcx(npy_cdouble z) { return to_ccomplex(xsf::erfcx(to_complex(z))); }
+
+double xsf_dawsn(double x) { return xsf::dawsn(x); }
+
+npy_cdouble xsf_cdawsn(npy_cdouble z) { return to_ccomplex(xsf::dawsn(to_complex(z))); }
+
+double xsf_erfi(double x) { return xsf::erfi(x); }
+
+npy_cdouble xsf_cerfi(npy_cdouble z) { return to_ccomplex(xsf::erfi(to_complex(z))); }
+
+npy_cdouble xsf_cwofz(npy_cdouble z) { return to_ccomplex(xsf::wofz(to_complex(z))); }
+
+double xsf_voigt_profile(double x, double sigma, double gamma) { return xsf::voigt_profile(x, sigma, gamma); }
 
 double cephes_ellpk(double x) { return xsf::ellipkm1(x); }
 
@@ -572,7 +592,7 @@ double xsf_gdtr(double a, double b, double x) { return xsf::gdtr(a, b, x); }
 
 double xsf_gdtrc(double a, double b, double x) { return xsf::gdtrc(a, b, x); }
 
-double xsf_gdtrib(double a, double p, double x) {return xsf::gdtrib(a, p, x); }
+double xsf_gdtrib(double a, double p, double x) { return xsf::gdtrib(a, p, x); }
 
 double xsf_kolmogorov(double x) { return xsf::kolmogorov(x); }
 
@@ -591,6 +611,12 @@ double xsf_nbdtrc(int k, int n, double p) { return xsf::nbdtrc(k, n, p); }
 double xsf_nbdtri(int k, int n, double p) { return xsf::nbdtri(k, n, p); }
 
 double xsf_ndtr(double x) { return xsf::ndtr(x); }
+
+npy_cdouble xsf_cndtr(npy_cdouble x) { return to_ccomplex(xsf::ndtr(to_complex(x))); }
+
+double xsf_log_ndtr(double x) { return xsf::log_ndtr(x); }
+
+npy_cdouble xsf_clog_ndtr(npy_cdouble x) { return to_ccomplex(xsf::log_ndtr(to_complex(x))); }
 
 double xsf_ndtri(double x) { return xsf::ndtri(x); }
